Delete Supplementary Data.xlsx
Thanks to Xiaoyang Zhong @ CML, Leiden Univesity, who noticed an error in the calculation of aluminium and glass in 3 residential building types. This commit fixes the problem and leads to somewhat lower glass and aluminium demand (in both stocks and flows) as described in the corrigendum submitted to the Journal of Cleaner Production. This commit removes the old Supplementary Data.xlsx